Move the implementation of `scoped` to `Child`
